refactor(testing): rearchitecture of how client is injected into CLI
